requirements.txt
streamlit
opencv-python
Pillow
numpy
scikit-learn 
tensorflow 